Added string handling and chaos
Added the ability to "calculate" strings. This works with all four basic operations, and they all do weird things.
Numbers are now simplified if possible.
Fixed a lot of bugs.
Drastically simplified most of the code and removed most redundancies.
Created a function file which holds all of the crazy calculator functions and simp(x) which simplifies basic numbers.